+++ examples/DeepSeekReviewFile.hs
@@ -66,7 +66,7 @@ main = do
                 , requestSource = source
                 }
         reviewer =
-            bind
+            mount
                 runtime
                 noTools
                 definition